Удалить контекст у TreeItem
